requirements.txt
# Base dependencies
numpy==1.24.3

# Deep learning
tensorflow==2.14.0
keras==2.14.0

# Image processing
Pillow>=11.1.0
opencv-python>=4.8.0

# Web framework and utilities
Flask>=3.1.0
python-dotenv>=1.1.0

# Face detection
--extra-index-url https://pypi.org/simple/
fire==0.7.0
deepface==0.0.93

# Additional dependencies for deepface
mtcnn>=0.1.0
retina-face>=0.0.1
gdown>=3.10.1
tqdm>=4.30.0
pandas>=0.23.4


